+++ editor/adf-schema/src/__tests__/unit/schema/marks/text-color.ts
@@ -20,7 +20,7 @@ describe(`${name}/schema textColor mark`, () => {
       schema.marks.textColor.create(testColorObj1),
     ]);
     expect(toHTML(node, schema)).toEqual(
-      `<span style="color: ${testColorObj1.color}">foo</span>`,
+      `<span class=\"fabric-text-color-mark\" style=\"--custom-text-color: ${testColorObj1.color}\" data-text-custom-color=\"#97a0af\">foo</span>`,
     );
   });
 
@@ -30,7 +30,7 @@ describe(`${name}/schema textColor mark`, () => {
       schema.marks.textColor.create(testColorObj2),
     ]);
     expect(toHTML(node, schema)).toEqual(
-      `<span style="color: ${testColorObj2.color}">foo</span>`,
+      `<span class=\"fabric-text-color-mark\" style=\"--custom-text-color: ${testColorObj2.color}\" data-text-custom-color=\"#97A0AF\">foo</span>`,
     );
   });
 });