+++ editor/editor-common/src/styles/shared/text-color.ts
@@ -1,0 +1,11 @@
+import { css } from 'styled-components';
+
+export const textColorStyles = css`
+  .fabric-text-color-mark {
+    color: var(--custom-text-color, inherit);
+  }
+
+  a .fabric-text-color-mark {
+    color: unset;
+  }
+`;