+++ editor/renderer/src/__tests__/unit/react/marks/textColor.tsx
@@ -1,21 +1,30 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { render, RenderResult } from '@testing-library/react';
 import TextColor from '../../../../react/marks/textColor';
 
 describe('Renderer - React/Marks/TextColor', () => {
-  const mark = shallow(
-    <TextColor dataAttributes={{ 'data-renderer-mark': true }} color="#ff0000">
-      This is a red text
-    </TextColor>,
-  );
-
-  it('should wrap content with <span>-tag', () => {
-    expect(mark.is('span')).toEqual(true);
+  let wrapper: RenderResult;
+  beforeEach(() => {
+    wrapper = render(
+      <TextColor
+        dataAttributes={{ 'data-renderer-mark': true }}
+        color="#ff0000"
+      >
+        This is a red text
+      </TextColor>,
+    );
   });
 
-  it('should output correct html', () => {
-    expect(mark.html()).toEqual(
-      '<span data-renderer-mark="true" style="color:#ff0000">This is a red text</span>',
+  it('should wrap content with <span>-tag', async () => {
+    const mark = await wrapper.getByText('This is a red text');
+    expect(mark.tagName).toEqual('SPAN');
+  });
+
+  it('should output correct html', async () => {
+    const mark = await wrapper.getByText('This is a red text');
+
+    expect(mark.outerHTML).toEqual(
+      `<span data-renderer-mark=\"true\" data-text-custom-color=\"#ff0000\" class=\"fabric-text-color-mark\" style=\"--custom-text-color: #ff0000;\">This is a red text</span>`,
     );
   });
 });