+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/editor-common
 
+## 62.1.0
+
+### Minor Changes
+
+- [`b9cd2373064`](https://bitbucket.org/atlassian/atlassian-frontend/commits/b9cd2373064) - [ED-14106] Reset custom color when the node is inside a link
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 62.0.0
 
 ### Major Changes